fix: vk model routing and logging plugin ticker fixes
