+++ plugins/governance/changelog.md
@@ -1,1 +1,2 @@
 - chore: upgrades core to 1.2.31 and framework to 1.1.40
+- feat: check allowed models from model catalog for provider configs
